requirements.txt
Mako==1.0.6
MarkupSafe==0.23
PasteDeploy==1.5.2
PyMySQL==0.7.11
Pygments==2.2.0
SQLAlchemy==1.1.9
sqlalchemy-utils==0.32.14
WebOb==1.7.1
alembic==0.9.1
passlib==1.7.1
PyExecJS==1.4.0
pylama==6.3.1
pyramid==1.8.3
pyramid-mako==1.0.2
pyramid-tm==0.11
repoze.lru==0.6
transaction==2.1.2
translationstring==1.3
venusian==1.0
waitress==1.0.2
zope.deprecation==4.2.0
zope.interface==4.4.0
zope.sqlalchemy==0.7.7
-e git://github.com/rickmak/pyramid_safile.git#egg=pyramid_safile
ply==3.6
rq==0.7.1
cornice==2.4.0
requests==2.14.2
stripe==1.55.2
google-cloud-storage==1.1.1
confluent-kafka==0.9.4
setuptools==35.0.2
firebase-admin==1.0.0
google-cloud-translate==0.24.0
Babel==2.5.0